feat(TV-57): relocate citizenapi test classes
